feat(tests): test default modules availability
We currently don't support specifying a precise version for these
modules. We just ship the latest version we have. So the test is rather
simple: just list available PHP modules and check we have everything.

The list of default modules will have to be maintained (or put somewhere
else?)

Also:
- refactor code in `test::helpers::test_compile`
- add an explanation message when skipping tests